MFT access in kernel lockdown via VFIO driver The purpose of this feature is to enable the livefish recovery flow to work in kernel lockdown mode, since user space access to the kernel is not allowed in that environment. Allow users to recover a Livefish device while the system is in kernel lockdown mode.

+++ include/mtcr_ul/mtcr_mf.h
@@ -120,7 +120,7 @@ struct mfile_t {
     unsigned int pxir_vsec_supp;
     u_int8_t       vsec_type;
     mtcr_status_e  icmd_support;
-    unsigned int   vsec_addr;
+    u_int64_t   vsec_addr;
     u_int32_t      vsec_cap_mask;
     int            address_space;
     int            multifunction;
